+++ tests/testthat/test_covid19br.R
@@ -1,5 +1,6 @@
 
 library(covid19br)
+library(dplyr)
 
 # Downloading Brazilian COVID-19 data:
 brazil <- downloadCovid19("brazil")
@@ -12,21 +13,25 @@ world <- downloadCovid19("world")
 
 
 
-# adding the geometry to the data:
+# adding the geometry/epidemiological rates to the downloaded data:
 regions_geo <- regions %>%
   filter(date == max(date)) %>%
-  add_geo()
+  add_geo() %>%
+  add_epi_rates()
 
 states_geo <- states %>%
   filter(date == max(date)) %>%
-  add_geo()
+  add_geo() %>%
+  add_epi_rates()
 
 cities_geo <- cities %>%
   filter(date == max(date)) %>%
-  add_geo()
+  add_geo() %>%
+  add_epi_rates()
 
 world_geo <- world %>%
   filter(date == max(date)) %>%
-  add_geo()
+  add_geo() %>%
+  add_epi_rates()